third_party/WebKit/LayoutTests/platform/chromium-linux/fast/forms/file/input-file-write-files-expected.txt

This tests the files attribute in file input forms

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS file1.files.length is 1
FAIL file1.files.item(0).name should be foo.txt. Was .
PASS file1.files.length is 1
FAIL file1.files.item(0).name should be foo.txt. Was .
PASS file1.files.length is 1
FAIL file1.files.item(0).name should be bar.txt. Was .
PASS successfullyParsed is true

TEST COMPLETE
 
